+++ lib/class/Provider.ts
@@ -5,15 +5,17 @@ import {
   NodeNameFilterType,
   ProviderConfig,
   SupportProviderEnum,
+  PossibleNodeConfigType,
+  SortedNodeNameFilterType,
 } from '../types';
 
 let globalPort: number = 61100;
 
 export default class Provider {
   public readonly type: SupportProviderEnum;
-  public readonly nodeFilter?: NodeFilterType;
-  public readonly netflixFilter?: NodeNameFilterType;
-  public readonly youtubePremiumFilter?: NodeNameFilterType;
+  public readonly nodeFilter?: ProviderConfig['nodeFilter'];
+  public readonly netflixFilter?: ProviderConfig['netflixFilter'];
+  public readonly youtubePremiumFilter?: ProviderConfig['youtubePremiumFilter'];
   public readonly customFilters?: ProviderConfig['customFilters'];
   public readonly addFlag?: boolean;
   public readonly tfo?: boolean;
@@ -24,10 +26,14 @@ export default class Provider {
       type: Joi.string()
         .valid(...Object.values<string>(SupportProviderEnum))
         .required(),
-      nodeFilter: Joi.function(),
-      netflixFilter: Joi.function(),
-      youtubePremiumFilter: Joi.function(),
-      customFilters: Joi.object().pattern(Joi.string(), Joi.function()),
+      nodeFilter: Joi.any().allow(Joi.function(), Joi.object({ filter: Joi.function(), supportSort: Joi.boolean() })),
+      netflixFilter: Joi.any().allow(Joi.function(), Joi.object({ filter: Joi.function(), supportSort: Joi.boolean() })),
+      youtubePremiumFilter: Joi.any().allow(Joi.function(), Joi.object({ filter: Joi.function(), supportSort: Joi.boolean() })),
+      customFilters: Joi.object()
+        .pattern(
+          Joi.string(),
+          Joi.any().allow(Joi.function(), Joi.object({ filter: Joi.function(), supportSort: Joi.boolean() }))
+        ),
       addFlag: Joi.boolean(),
       startPort: Joi.number().integer().min(1024).max(65535),
     })
@@ -56,4 +62,8 @@ export default class Provider {
     }
     return globalPort++;
   }
+
+  public getNodeList(): Promise<ReadonlyArray<PossibleNodeConfigType>> {
+    return Promise.resolve([]);
+  };
 }