+++ test/fixture/custom-filter/provider/ss.js
@@ -2,11 +2,14 @@
 
 const { utils } = require('../../../../');
 
+exports.keywordFilter = utils.useKeywords(['US 1', 'US 2']);
+exports.strictKeywordFilter = utils.useKeywords(['US', 'Netflix'], true);
+
 module.exports = {
   url: 'http://example.com/test-ss-sub.txt',
   type: 'shadowsocks_subscribe',
   customFilters: {
-    keywordFilter: utils.useKeywords(['US 1', 'US 2']),
-    strictKeywordFilter: utils.useKeywords(['US', 'Netflix'], true),
+    keywordFilter: exports.keywordFilter,
+    strictKeywordFilter: exports.strictKeywordFilter,
   },
 };